refactor: relocate device (API)
