+++ src/connectivity/__mocks__/events.ts
@@ -1,4 +1,4 @@
-import { DeviceType } from "../device-info";
+import { DeviceType } from "../../api/device";
 import type * as events from "../events";
 
 const action = "com.elgato.test.one";
